💡 Add: 스타일 파일 분리 및 컴포넌트  주석 추가(#3)
