+++ src/routes/activity/components/YearTabs.jsx
@@ -1,26 +1,43 @@
-// YearTabs.jsx
-import React from 'react';
-import styled from 'styled-components';
-import palette from '../../../styles/theme'; // 경로 프로젝트에 맞게
+/**
+ * YearTabs
+ * 연도 탭 네비게이션 컴포넌트
+ *
+ * 기능
+ * - 연도 리스트를 탭으로 노출, 클릭/키보드(←/→)로 선택 가능
+ * - controlled(value/onChange) & uncontrolled(defaultValue) 모두 지원
+ * - 비활성 연도(disabledYears) 처리
+ *
+ * Props
+ * - years {number[]} 탭으로 보여줄 연도 배열 (미지정 시 현재연도~+2)
+ * - value {number} 선택된 연도(컨트롤드)
+ * - defaultValue {number} 초기 선택값(언컨트롤드)
+ * - onChange {(year:number)=>void} 선택 변경 콜백
+ * - gap {number} 탭 사이 간격(px), 기본 32
+ * - size {'md'|'lg'} 탭 폰트 크기 프리셋, 기본 'lg'
+ * - disabledYears {number[]} 비활성 처리할 연도들
+ * - className {string} 커스텀 클래스
+ *
+ * 사용 예시
+ * <YearTabs years={[2025,2026,2027]} value={year} onChange={setYear} size="lg" />
+ */
 
-const NAVY = (palette && palette.mainNavy?.navy100) || '#1f2a68';
-const GRAY = (palette && palette.text?.tertiary) || '#c7cbe0';
+import React from 'react';
+import * as S from '@routes/activity/components/YearTabsStyle';
 
 export default function YearTabs({
   years,
-  value, // controlled: 선택된 연도
-  defaultValue, // uncontrolled 초기값
-  onChange, // (year) => void
-  gap = 32, // 탭 간격(px)
-  size = 'lg', // 'md' | 'lg'
-  disabledYears = [], // 비활성 연도 리스트
+  value,
+  defaultValue,
+  onChange,
+  gap = 32,
+  size = 'lg',
+  disabledYears = [],
   className,
 }) {
   const fallbackYears = React.useMemo(() => {
     const y = new Date().getFullYear();
     return [y, y + 1, y + 2];
   }, []);
-
   const list = years && years.length ? years : fallbackYears;
 
   const [internal, setInternal] = React.useState(defaultValue ?? list[0]);
@@ -42,12 +59,12 @@ export default function YearTabs({
   };
 
   return (
-    <Wrap role="tablist" aria-label="연도 선택" $gap={gap} className={className} onKeyDown={onKeyDown}>
+    <S.Wrap role="tablist" aria-label="연도 선택" $gap={gap} className={className} onKeyDown={onKeyDown}>
       {list.map((y) => {
         const disabled = disabledYears.includes(y);
         const active = String(selected) === String(y);
         return (
-          <YearBtn
+          <S.YearBtn
             key={y}
             role="tab"
             aria-selected={active}
@@ -58,66 +75,9 @@ export default function YearTabs({
             data-disabled={disabled || undefined}
             onClick={() => !disabled && setSelected(y)}>
             {y}
-          </YearBtn>
+          </S.YearBtn>
         );
       })}
-    </Wrap>
+    </S.Wrap>
   );
 }
-
-/* ------------ styles ------------ */
-const Wrap = styled.div`
-  display: flex;
-  align-items: flex-end;
-  justify-content: center;
-  gap: ${(p) => p.$gap}px;
-`;
-
-const YearBtn = styled.button`
-  appearance: none;
-  background: transparent;
-  border: 0;
-  padding: 0 2px 8px; /* 밑줄 여유 */
-  margin: 0;
-  cursor: pointer;
-  position: relative;
-
-  color: ${GRAY};
-  font-weight: 700;
-  line-height: 1;
-
-  /* 사이즈 */
-  &[data-size='lg'] {
-    font-size: 14px;
-  }
-  &[data-size='md'] {
-    font-size: 14px;
-  }
-
-  /* 활성 */
-  &[data-active] {
-    color: ${NAVY};
-  }
-  &[data-active]::after {
-    content: '';
-    position: absolute;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    height: 2px;
-    background: ${NAVY};
-    border-radius: 2px;
-  }
-
-  /* 비활성 */
-  &[data-disabled] {
-    opacity: 0.45;
-    cursor: not-allowed;
-  }
-
-  /* 접근성 포커스 */
-  &:focus-visible {
-    outline: none;
-    box-shadow: inset 0 -2px 0 ${NAVY};
-  }
-`;
